Change the create service version endpoint
We have updated the create service version endpoint to be a PUT to the
url `services/<service_id>/versions` as we felt this was a better
representation of what the Metadata API is actually doing. It is
creating a brand new version of the metadata in the DB.

We also removed some of the 422 respsones from the GET requests.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>